quickstart-python/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
attrs==22.1.0
    # via nile-api
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
emoji==2.1.0
    # via -r requirements.in
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via nile-api
idna==3.4
    # via
    #   anyio
    #   rfc3986
nile-api==0.2.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via nile-api
python-dotenv==0.21.0
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
